backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# Database drivers (async)
asyncpg==0.29.0  # PostgreSQL
aiomysql==0.2.0  # MySQL
aioodbc==0.5.0  # SQL Server
aiosqlite==0.19.0  # SQLite

# Excel processing
pandas==2.1.3
openpyxl==3.1.2
xlrd==2.0.1

# Validation and settings
pydantic==2.5.0
pydantic-settings==2.1.0

# Authentication
python-jose[cryptography]
passlib[bcrypt]

# Environment variables
python-dotenv==1.0.0

# CORS is included in FastAPI, no separate package needed

# Utilities
python-dateutil==2.8.2

# Logging
structlog==23.2.0

